copy_host_files: Support different users depending on lab chosen
This patch adds support for users to override the user used to connect
to remote hosts.  By default, the 'ubuntu' user will be used, as this is
the one associated with the 'docker' lab used by train.  Else the user
can override this to anything they choose, this should commonly be used
with environments like 'base'.

Signed-off-by: Kyle Squizzato <kyle.squizzato@docker.com>